Verifying that autoplay=true sets off the error
i.e. Mixed content block happens when it tries to ambush the player, basically
Even though this didn't seem to be the case for the non-positional, but that's probably because it was inside of <audio>, an HTML5 element ("passive content" to the browser) whereas <a-sound> for positional SFX runs through A-Frame's <script> (potentially malicious "active mixed content").